admin/requirements.txt
Django==4.1
djangorestframework==3.12.2
mysqlclient==2.2.4
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
pytz==2024.2
python-dotenv==1.0.1
requests==2.32.0